requirements.txt
Django==3.2.16
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-solo==2.0.0
djangorestframework==3.14.0